+++ crates/execution/ab-riscv-primitives/src/instruction/b_64_ext/zbs_64_ext.rs
@@ -82,8 +82,8 @@ where
     }
 
     #[inline(always)]
-    fn size(&self) -> usize {
-        size_of::<u32>()
+    fn size(&self) -> u8 {
+        size_of::<u32>() as u8
     }
 }
