+++ crates/execution/ab-riscv-primitives/src/instruction/m_64_ext.rs
@@ -87,8 +87,8 @@ where
     }
 
     #[inline(always)]
-    fn size(&self) -> usize {
-        size_of::<u32>()
+    fn size(&self) -> u8 {
+        size_of::<u32>() as u8
     }
 }
